requirements.txt
Django==3.2.6
Pillow==8.4.0
python-decouple==3.5
